Move analyzer package logic into brokk-shared
