--- app/src/main/java/ai/brokk/analyzer/package-info.java
@@ -1,4 +1,0 @@
-@NullMarked
-package ai.brokk.analyzer;
-
-import org.jspecify.annotations.NullMarked;
