+++ app/src/main/java/ai/brokk/context/ContextFragments.java
@@ -17,12 +17,12 @@ import ai.brokk.analyzer.MultiAnalyzer;
 import ai.brokk.analyzer.ProjectFile;
 import ai.brokk.analyzer.TypeHierarchyProvider;
 import ai.brokk.analyzer.usages.FuzzyResult;
-import ai.brokk.analyzer.usages.UsageFinder;
 import ai.brokk.analyzer.usages.UsageHit;
 import ai.brokk.concurrent.ComputedValue;
 import ai.brokk.concurrent.ExecutorsUtil;
 import ai.brokk.concurrent.LoggingExecutorService;
 import ai.brokk.git.GitRepo;
+import ai.brokk.usages.UsageFinder;
 import ai.brokk.util.FragmentUtils;
 import ai.brokk.util.ImageUtil;
 import ai.brokk.util.Lines;
