+++ app/src/main/java/ai/brokk/usages/package-info.java
@@ -1,4 +1,4 @@
 @NullMarked
-package ai.brokk.analyzer.usages;
+package ai.brokk.usages;
 
 import org.jspecify.annotations.NullMarked;
