+++ app/src/test/java/ai/brokk/tools/UsageBenchEval.java
@@ -6,9 +6,7 @@ import ai.brokk.analyzer.IAnalyzer;
 import ai.brokk.analyzer.Language;
 import ai.brokk.analyzer.Languages;
 import ai.brokk.analyzer.ProjectFile;
-import ai.brokk.analyzer.usages.LlmUsageAnalyzer;
 import ai.brokk.analyzer.usages.UsageAnalyzer;
-import ai.brokk.analyzer.usages.UsageFinder;
 import ai.brokk.analyzer.usages.UsageHit;
 import ai.brokk.concurrent.ExecutorsUtil;
 import ai.brokk.concurrent.LoggingExecutorService;
@@ -16,6 +14,8 @@ import ai.brokk.project.IProject;
 import ai.brokk.project.ModelProperties;
 import ai.brokk.testutil.TestProject;
 import ai.brokk.tools.UsageBenchTypes.*;
+import ai.brokk.usages.LlmUsageAnalyzer;
+import ai.brokk.usages.UsageFinder;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.nio.file.Files;
