+++ brokk-shared/src/main/java/ai/brokk/analyzer/AGENTS.md
@@ -2,7 +2,7 @@
 
 ### 1. Prefer Tree-sitter Over Regex
 1. **Tree-sitter queries**: Use Tree-sitter queries (`.scm` files) for parsing structured code; they handle whitespace, comments, and edge cases that regex cannot.
-1. **Regex usage**: When regex IS the better solution (e.g., simple string patterns, non-AST text processing), always compile patterns as `private static final Pattern` fields — never inline `Pattern.compile()` calls.
+1. **Regex usage**: When regex IS the better solution (e.g., simple string patterns, non-AST text processing), always compile patterns as `private static final Pattern` fields - never inline `Pattern.compile()` calls.
    - Example:
      ```java
      private static final Pattern WILDCARD_IMPORT_PATTERN = Pattern.compile("^from\\s+(.+?)\\s+import\\s+\\*");
@@ -14,12 +14,12 @@
 
 ### 3. Working with SourceContent
 1. **Creation**: Use `SourceContent.of(String source)` to create a content wrapper that handles UTF-8 byte offset conversions.
-1. **Extraction**: Extract text from AST nodes using `sourceContent.substringFrom(TSNode node)` — this correctly handles multi-byte Unicode characters.
+1. **Extraction**: Extract text from AST nodes using `sourceContent.substringFrom(TSNode node)` - this correctly handles multi-byte Unicode characters.
 1. **Avoid manual calculations**: Avoid manual byte-to-char offset calculations; the `SourceContent` helper methods (`substringFromBytes`, `byteOffsetToCharPosition`, `charPositionToByteOffset`) handle edge cases.
 
 ### 4. Analyzer Snapshot Architecture
-1. **Immutability**: Analyzers are **immutable snapshots** of the project state at a point in time.
-1. **Update behavior**: Calling `analyzer.update()` returns a **new** analyzer instance; it does not mutate the existing one. Instance fields on an analyzer represent the state of that specific snapshot and should not be modified.
+1. **Immutability**: Analyzers are immutable snapshots of the project state at a point in time.
+1. **Update behavior**: Calling `analyzer.update()` returns a new analyzer instance; it does not mutate the existing one. Instance fields on an analyzer represent the state of that specific snapshot and should not be modified.
 1. **Persistent State**: If state needs to persist across `update()` calls (e.g., caches, indexes), it must be stored in the `AnalyzerState` record.
 1. **PMap usage**: `AnalyzerState` uses `PMap` (persistent/immutable maps from `pcollections`) to enable efficient structural sharing and avoid unnecessary garbage collection.
 1. **Lazy caches**: Lazy caches (like `LazyTypeHierarchyCache`, `LazyImportCache`) are instance-level and reset on `update()`; their computed values are merged into `AnalyzerState` via `snapshotState()` for serialization.
@@ -33,9 +33,9 @@
    - `resources/treesitter/<language>/definitions.scm`
    - `resources/treesitter/<language>/imports.scm`
 1. **QueryType Enum**:
-   - `DEFINITIONS`: Primary query for capturing classes, functions, and fields. This is **compulsory** for every analyzer.
-   - `IMPORTS`: Specific query for capturing import statements and related symbols. This is **optional**; if not provided, the analyzer will skip import-specific AST passes.
-   - `IDENTIFIERS`: Specific query for capturing type identifiers and references. This is **optional**; used for advanced symbol resolution.
+   - `DEFINITIONS`: Primary query for capturing classes, functions, and fields. This is compulsory for every analyzer.
+   - `IMPORTS`: Specific query for capturing import statements and related symbols. This is optional; if not provided, the analyzer will skip import-specific AST passes.
+   - `IDENTIFIERS`: Specific query for capturing type identifiers and references. This is optional; used for advanced symbol resolution.
 1. **Resource Loading**: The `TreeSitterAnalyzer` base class handles the discovery and loading of these queries based on the language name and `QueryType`. Analyzers should transition away from the single `getQueryResource()` string toward this structured multi-query approach.
 
 ### Tree-sitter Generated Constants (Preferred)
@@ -76,3 +76,4 @@ TSNode nameNode = methodNode.getChildByFieldName(nodeField(JavaNodeField.NAME));
 1. When editing analyzers, replace string node type checks (`"foo_bar".equals(node.getType())`) with the generated `*NodeType` / `*NodeField` constants where available.
 2. If the language does not have generated schema/constants yet, use the existing `*TreeSitterNodeTypes` / `CommonTreeSitterNodeTypes` constants as a fallback (still avoid raw strings).
 3. Query capture names (the names in `.scm` files like `import.declaration` or `test_marker`) are not node types. Keep those as explicit string constants (e.g., `IMPORT_DECLARATION_CAPTURE`), do not try to model them as `NodeType`s.
+
